revm/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via jsonschema
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
cryptography==36.0.1
    # via pyjwt
defusedxml==0.7.1
    # via
    #   odfpy
    #   python3-openid
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
dj-rest-auth[with_social]==2.2.3
    # via -r requirements.in
django==3.2.12
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   django-import-export
    #   django-jazzmin
    #   django-multiselectfield
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==0.47.0
    # via dj-rest-auth
django-cors-headers==3.11.0
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-import-export==2.7.1
    # via -r requirements.in
django-jazzmin==2.4.9
    # via -r requirements.in
django-multiselectfield==0.1.12
    # via -r requirements.in
django-storages==1.12.3
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.0.0
    # via -r requirements.in
drf-spectacular==0.21.2
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gevent==21.12.0
    # via -r requirements.in
greenlet==1.1.2
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.4.0
    # via drf-spectacular
markuppy==1.14
    # via tablib
oauthlib==3.2.0
    # via requests-oauthlib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
psycopg2==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
pymemcache==3.5.1
    # via -r requirements.in
pyrsistent==0.18.1
    # via jsonschema
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via
    #   django
    #   djangorestframework
pyyaml==6.0
    # via
    #   drf-spectacular
    #   tablib
requests==2.27.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via pymemcache
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.0
    # via django-import-export
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.8
    # via requests
whitenoise==6.0.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
